rename TableDefinition to BaseTable
